+++ src/test/kotlin/christmas/domain/promotion/StarDiscountTest.kt
@@ -1,5 +1,6 @@
 package christmas.domain.promotion
 
+import christmas.configuration.promotion.StarDiscount
 import christmas.domain.money.Currency
 import christmas.domain.money.Money
 import christmas.domain.order.Orders
